+++ opencontractserver/constants/__init__.py
@@ -9,5 +9,6 @@ from opencontractserver.constants.auth import *  # noqa: F401, F403
 from opencontractserver.constants.context_guardrails import *  # noqa: F401, F403
 from opencontractserver.constants.corpus_actions import *  # noqa: F401, F403
 from opencontractserver.constants.document_processing import *  # noqa: F401, F403
+from opencontractserver.constants.llm_tools import *  # noqa: F401, F403
 from opencontractserver.constants.moderation import *  # noqa: F401, F403
 from opencontractserver.constants.zip_import import *  # noqa: F401, F403